cast double pr eviter troncature temps execution

+++ src/main/java/fr/abes/bestppn/utils/ExecutionTimeAspect.java
@@ -16,7 +16,7 @@ public class ExecutionTimeAspect {
         Object result = joinPoint.proceed();
 
         long endTime = System.currentTimeMillis();
-        double executionTime = (endTime - startTime) / 1000;
+        double executionTime = (double) (endTime - startTime) / 1000;
 
         log.debug("Temps d'exécution : " + executionTime + " secondes");
         return result;
